fix: Address code review findings - type consistency and null safety
- Change ImpersonationAlertRecord.Id from int to long to match database schema
- Add defensive null check after re-fetching exam session before evaluation

+++ TelegramGroupsAdmin.Core/Models/ImpersonationAlertRecord.cs
@@ -8,7 +8,7 @@ namespace TelegramGroupsAdmin.Core.Models;
 /// </summary>
 public record ImpersonationAlertRecord
 {
-    public int Id { get; init; }
+    public long Id { get; init; }
     public long SuspectedUserId { get; init; }
     public long TargetUserId { get; init; }
     public long ChatId { get; init; }
